Changed opensource to free and open source
In my opinion it is important to distinguish that we want to support
Free and Open Source software, and not only open source. The term
'open source' is often misused for software where one can look at the
source code under some NDA. Free Software makes sure, that the
specific software has a DFSG-compatible license.