Modify URL handling in LaTeX (partial fix for #55)
Specifically, we add the url package, as well as a small sed script to
rewrite any hrefs with identical link and text to be a \url instead. This
fixes the majority of the text-related hbox overruns described in #55

NOTE: the sed script is stupid and assumes that if the link text starts
with "http" then it's the same as the link destination.